modularize `scheduling/shifts` routes into controllers directory
